+++ lib/pio/open_flow10.rb
@@ -5,6 +5,7 @@ module Pio
   end
 end
 
+require 'pio/open_flow10/actions'
 require 'pio/open_flow10/echo'
 require 'pio/open_flow10/exact_match'
 require 'pio/open_flow10/features'