+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## develop (unreleased)
 
 
+## 0.21.0 (6/19/2015)
+### New features
+* [#164](https://github.com/trema/pio/pull/164): Add new classes `Pio::FlowMod`, `Pio::Apply`, `Pio::GotoTable`, `Pio::WriteMetadata`, `Pio::Meter` and `Pio::SendOutPort`.
+
+
 ## 0.20.1 (6/10/2015)
 ### Bugs fixed
 * [#167](https://github.com/trema/pio/pull/167): Fix PacketIn accessor methods (raw_data = VLAN tagged UDP).